Moved SendAttribute call to GMHandler. Added mode verification. Reduced repetition.
