+++ src/Imgeneus.Network/Packets/Game/CharacterAttributeEnum.cs
@@ -1,0 +1,25 @@
+namespace Imgeneus.Network.Packets.Game
+{
+    public enum CharacterAttributeEnum : byte
+    {
+        Grow = 0,
+        Level = 1,
+        Money = 2,
+        StatPoint = 3,
+        SkillPoint = 4,
+        Strength = 5,
+        Dexterity = 6,
+        Reaction = 7,
+        Intelligence = 8,
+        Luck = 9,
+        Wisdom = 10,
+        Hg = 11, // ??
+        Vg = 12, // ??
+        Cg = 13, // ??
+        Og = 14, // ??
+        Ig = 15, // ??
+        Exp = 16,
+        Kills = 17,
+        Deaths = 18
+    }
+}
